requirements.txt

# Workshop App Requirements
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.5.0
requests>=2.31.0
zxcvbn>=4.4.28
yfinance>=0.2.18
python-dateutil>=2.8.2
pytz>=2023.3
